+++ lib/src/surveys/models/posthog_display_choice_question.dart
@@ -6,6 +6,7 @@ import 'posthog_display_survey_question.dart';
 @immutable
 class PostHogDisplayChoiceQuestion extends PostHogDisplaySurveyQuestion {
   const PostHogDisplayChoiceQuestion({
+    required super.id,
     required super.question,
     required this.choices,
     required this.isMultipleChoice,